+++ server/src/main/java/nc/students/ncvito/service/AnnouncementService.java
@@ -1,10 +1,7 @@
 package nc.students.ncvito.service;
 
 
-import nc.students.ncvito.entity.Announcement;
-import nc.students.ncvito.entity.Favorites;
-import nc.students.ncvito.entity.Status;
-import nc.students.ncvito.entity.User;
+import nc.students.ncvito.entity.*;
 import nc.students.ncvito.repo.AnnouncementRepository;
 import nc.students.ncvito.repo.FavoritesRepository;
 import nc.students.ncvito.repo.UserRepository;
@@ -14,7 +11,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.Collections;
+import java.util.List;
 
 
 @Service
@@ -38,13 +37,30 @@ public class AnnouncementService {
         return announcementRepository.findById(id);
     }
 
-    public void delete(Announcement announcement) {
-        announcementRepository.delete(announcement);
+    public void delete(Announcement announcement, Authentication authentication) {
+
+        User user = userRepository.findByLogin(authentication.getName())
+                .orElseThrow(() -> new IllegalStateException("User with login " + authentication.getName() + " not found."));
+
+        if (user.getId() == announcement.getAuthor().getId() || user.getRole().contains(Role.ADMIN) || user.getRole().contains(Role.MODERATOR)) {
+            List<Favorites> favoritesList = favoritesRepository.findByAnnouncement(announcement);
+            for (Favorites favorite : favoritesList) {
+                favoritesRepository.delete(favorite);
+            }
+            announcementRepository.delete(announcement);
+        }
+
+
     }
 
 
-    public Announcement update(Announcement announcement) {
-        return announcementRepository.save(announcement);
+    public Announcement update(Announcement announcement, Authentication authentication) {
+        User user = userRepository.findByLogin(authentication.getName())
+                .orElseThrow(() -> new IllegalStateException("User with login " + authentication.getName() + " not found."));
+
+        if (user.getId() == announcement.getAuthor().getId() || user.getRole().contains(Role.ADMIN) || user.getRole().contains(Role.MODERATOR)) {
+            return announcementRepository.save(announcement);
+        } else return null;
     }
 
     public Announcement create(Announcement announcement, Authentication authentication) {
@@ -63,6 +79,27 @@ public class AnnouncementService {
         favorites.setUser(user);
         favoritesRepository.save(favorites);
 
+
+    }
+
+    public void deleteFavorites(Announcement announcement, Authentication authentication) {
+        User user = userRepository.findByLogin(authentication.getName())
+                .orElseThrow(() -> new IllegalStateException("User with login " + authentication.getName() + " not found."));
+        Favorites favorites = favoritesRepository.findByAnnouncementAndUser(announcement, user);
+        favoritesRepository.delete(favorites);
+
+    }
+
+    public List<Announcement> getAllFavoritesByUser(Authentication authentication) {
+        User user = userRepository.findByLogin(authentication.getName())
+                .orElseThrow(() -> new IllegalStateException("User with login " + authentication.getName() + " not found."));
+        List<Announcement> announcementList = new ArrayList<>();
+        List<Favorites> favoritesList = favoritesRepository.findAllByUser(user);
+        for (Favorites favorites : favoritesList) {
+            announcementList.add(favorites.getAnnouncement());
+        }
+
+        return announcementList;
 
     }
 }
